+++ authlete/dto/device_verification_response.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -31,7 +31,8 @@ class DeviceVerificationResponse(ApiResponse):
             'scopes':            Scope, # list of Scope
             'claimNames':        str,   # list of str
             'acrs':              str,   # list of str
-            'expiresAt':         int
+            'expiresAt':         int,
+            'resources':         str,   # list of str
         }
 
         super().__init__(nameAndValues, nameAndTypes)