+++ CHANGES.md
@@ -37,6 +37,9 @@ CHANGES
 - `DeviceAuthorizationResponse` class
     * Added `resources`.
 
+- `DeviceVerificationResponse` class
+    * Added `resources`.
+
 
 1.0.2 (2019-08-09)
 ------------------